Relocate context theme, rename module and exports
