+++ core/App/components/record/RecordFooter.tsx
@@ -1,10 +1,10 @@
 import React from 'react'
 import { StyleSheet, View } from 'react-native'
 
-import { useThemeContext } from '../../utils/themeContext'
+import { useTheme } from '../../contexts/theme'
 
 const RecordFooter: React.FC = ({ children }) => {
-  const { ColorPallet } = useThemeContext()
+  const { ColorPallet } = useTheme()
   const styles = StyleSheet.create({
     container: {
       backgroundColor: ColorPallet.brand.secondaryBackground,
