+++ core/App/index.ts
@@ -22,7 +22,7 @@ export { StoreProvider, StoreContext } from './contexts/store'
 export type { StoreProviderProps } from './contexts/store'
 export { theme } from './theme'
 export type { Theme } from './theme'
-export { ThemeProvider, useThemeContext } from './utils/themeContext'
+export { ThemeProvider, useTheme } from './contexts/theme'
 export { ConfigurationProvider, useConfiguration } from './contexts/configuration'
 export type { ConfigurationContext } from './contexts/configuration'
 export { default as defaultTerms } from './screens/Terms'
