+++ core/App/navigators/ConnectStack.tsx
@@ -1,15 +1,15 @@
 import { createStackNavigator } from '@react-navigation/stack'
 import React from 'react'
 
+import { useTheme } from '../contexts/theme'
 import Scan from '../screens/Scan'
 import { ConnectStackParams, Screens } from '../types/navigators'
-import { useThemeContext } from '../utils/themeContext'
 
 import { createDefaultStackOptions } from './defaultStackOptions'
 
 const ConnectStack: React.FC = () => {
   const Stack = createStackNavigator<ConnectStackParams>()
-  const theme = useThemeContext()
+  const theme = useTheme()
   const defaultStackOptions = createDefaultStackOptions(theme)
   return (
     <Stack.Navigator screenOptions={{ ...defaultStackOptions, headerShown: false }}>
