+++ core/App/navigators/SettingStack.tsx
@@ -2,17 +2,17 @@ import { createStackNavigator } from '@react-navigation/stack'
 import React from 'react'
 import { useTranslation } from 'react-i18next'
 
+import { useTheme } from '../contexts/theme'
 import Language from '../screens/Language'
 import Settings from '../screens/Settings'
 import { Screens, SettingStackParams } from '../types/navigators'
-import { useThemeContext } from '../utils/themeContext'
 
 import { createDefaultStackOptions } from './defaultStackOptions'
 
 const SettingStack: React.FC = () => {
   const Stack = createStackNavigator<SettingStackParams>()
   const { t } = useTranslation()
-  const theme = useThemeContext()
+  const theme = useTheme()
   const defaultStackOptions = createDefaultStackOptions(theme)
   return (
     <Stack.Navigator screenOptions={{ ...defaultStackOptions }}>
