+++ core/App/navigators/TabStack.tsx
@@ -5,10 +5,10 @@ import { Text, View } from 'react-native'
 import { SafeAreaView } from 'react-native-safe-area-context'
 import Icon from 'react-native-vector-icons/MaterialCommunityIcons'
 
+import { useTheme } from '../contexts/theme'
 import { useNotifications } from '../hooks/notifications'
 import { ColorPallet, TextTheme } from '../theme'
 import { Screens, Stacks, TabStackParams, TabStacks } from '../types/navigators'
-import { useThemeContext } from '../utils/themeContext'
 
 import CredentialStack from './CredentialStack'
 import HomeStack from './HomeStack'
@@ -17,7 +17,7 @@ const TabStack: React.FC = () => {
   const { total } = useNotifications()
   const { t } = useTranslation()
   const Tab = createBottomTabNavigator<TabStackParams>()
-  const { ColorPallet, TextTheme } = useThemeContext()
+  const { ColorPallet, TextTheme } = useTheme()
   return (
     <SafeAreaView style={{ flex: 1, backgroundColor: ColorPallet.brand.primary }}>
       <Tab.Navigator
