+++ core/__tests__/components/AvatarView.test.tsx
@@ -2,12 +2,12 @@ import { render } from '@testing-library/react-native'
 import React from 'react'
 
 import AvatarView from '../../App/components/misc/AvatarView'
+import * as themeContext from '../../App/contexts/theme' // note we're importing with a * to import all the exports
 import { theme } from '../../App/theme'
-import * as themeContext from '../../App/utils/themeContext' // note we're importing with a * to import all the exports
 
 describe('AvatarView Component', () => {
   test('Renders correctly', async () => {
-    jest.spyOn(themeContext, 'useThemeContext').mockImplementation(() => theme)
+    jest.spyOn(themeContext, 'useTheme').mockImplementation(() => theme)
     const tree = render(<AvatarView name={'Bacon'} />)
 
     expect(tree).toMatchSnapshot()
