+++ core/__tests__/components/ErrorModal.test.tsx
@@ -2,8 +2,8 @@ import { render, fireEvent } from '@testing-library/react-native'
 import React, { useContext } from 'react'
 
 import ErrorModal from '../../App/components/modals/ErrorModal'
+import * as themeContext from '../../App/contexts/theme' // note we're importing with a * to import all the exports
 import { theme } from '../../App/theme'
-import * as themeContext from '../../App/utils/themeContext' // note we're importing with a * to import all the exports
 
 jest.mock('react', () => ({
   ...jest.requireActual('react'),
@@ -28,7 +28,7 @@ useContext.mockImplementation(() => [state, setState])
 
 describe('ErrorModal Component', () => {
   test('Renders correctly', async () => {
-    jest.spyOn(themeContext, 'useThemeContext').mockImplementation(() => theme)
+    jest.spyOn(themeContext, 'useTheme').mockImplementation(() => theme)
     const tree = render(<ErrorModal />)
 
     expect(tree).toMatchSnapshot()
@@ -36,7 +36,7 @@ describe('ErrorModal Component', () => {
   })
 
   test('Dismiss on demand', async () => {
-    jest.spyOn(themeContext, 'useThemeContext').mockImplementation(() => theme)
+    jest.spyOn(themeContext, 'useTheme').mockImplementation(() => theme)
     const tree = render(<ErrorModal />)
 
     const dismissBtn = await tree.findByText('Global.Okay')
