+++ core/__tests__/screens/Onboarding.test.tsx
@@ -2,10 +2,10 @@ import { render } from '@testing-library/react-native'
 import React from 'react'
 import { Text } from 'react-native'
 
+import * as themeContext from '../../App/contexts/theme' // note we're importing with a * to import all the exports
 import Onboarding, { OnboardingStyleSheet } from '../../App/screens/Onboarding'
 import { createCarouselStyle } from '../../App/screens/OnboardingPages'
 import { theme } from '../../App/theme'
-import * as themeContext from '../../App/utils/themeContext' // note we're importing with a * to import all the exports
 
 export const carousel: OnboardingStyleSheet = createCarouselStyle(theme)
 
@@ -20,7 +20,7 @@ const pages = [
 
 describe('Onboarding', () => {
   it('Renders correctly', () => {
-    jest.spyOn(themeContext, 'useThemeContext').mockImplementation(() => theme)
+    jest.spyOn(themeContext, 'useTheme').mockImplementation(() => theme)
     const tree = render(<Onboarding pages={pages} nextButtonText="Next" previousButtonText="Back" style={carousel} />)
 
     expect(tree).toMatchSnapshot()
@@ -35,7 +35,7 @@ describe('Onboarding', () => {
   // })
 
   it('Pages exist', async () => {
-    jest.spyOn(themeContext, 'useThemeContext').mockImplementation(() => theme)
+    jest.spyOn(themeContext, 'useTheme').mockImplementation(() => theme)
     const { findAllByTestId } = render(
       <Onboarding pages={pages} nextButtonText="Next" previousButtonText="Back" style={carousel} />
     )
